+++ easypermissions-ktx/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <!--suppress AndroidDomInspection -->
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mvalceleanu.easypermissions">
+    package="com.vmadalin.easypermissions">
 
     <application>
         <activity android:name=".components.TestActivity"/>
