+++ easypermissions-ktx/src/main/java/com/vmadalin/easypermissions/annotations/AfterPermissionGranted.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mvalceleanu.easypermissions.annotations
+package com.vmadalin.easypermissions.annotations
 
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.FUNCTION)
