+++ easypermissions-ktx/src/main/java/com/vmadalin/easypermissions/utils/AnnotationsUtils.kt
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mvalceleanu.easypermissions.utils
+package com.vmadalin.easypermissions.utils
 
 import android.util.Log
-import com.mvalceleanu.easypermissions.annotations.AfterPermissionGranted
+import com.vmadalin.easypermissions.annotations.AfterPermissionGranted
 import java.lang.reflect.InvocationTargetException
 import kotlin.reflect.KClass
