+++ easypermissions-ktx/src/test/java/pub/devrel/easypermissions/annotations/Mockable.kt
@@ -13,6 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mvalceleanu.easypermissions.annotations
+package com.vmadalin.easypermissions.annotations
 
 annotation class Mockable
