+++ easypermissions-ktx/src/test/java/pub/devrel/easypermissions/components/TestFragment.kt
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mvalceleanu.easypermissions.components
+package com.vmadalin.easypermissions.components
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.mvalceleanu.easypermissions.R
-import com.mvalceleanu.easypermissions.annotations.AfterPermissionGranted
-import com.mvalceleanu.easypermissions.annotations.Mockable
-import com.mvalceleanu.easypermissions.EasyPermissions
+import com.vmadalin.easypermissions.R
+import com.vmadalin.easypermissions.annotations.AfterPermissionGranted
+import com.vmadalin.easypermissions.annotations.Mockable
+import com.vmadalin.easypermissions.EasyPermissions
 
 @Mockable
 class TestFragment :
