+++ easypermissions-ktx/src/test/java/pub/devrel/easypermissions/components/TestSupportFragmentActivity.kt
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mvalceleanu.easypermissions.components
+package com.vmadalin.easypermissions.components
 
 import androidx.fragment.app.FragmentActivity
-import com.mvalceleanu.easypermissions.annotations.Mockable
-import com.mvalceleanu.easypermissions.EasyPermissions
+import com.vmadalin.easypermissions.annotations.Mockable
+import com.vmadalin.easypermissions.EasyPermissions
 
 @Mockable
 class TestSupportFragmentActivity :
